Fixed data type to long double

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -11,21 +11,21 @@
 int main(void)
 {
 	int i = 0, m = 3;
-	unsigned long int j = 1, k = 2;
+	long double j = 1, k = 2;
 
 	while (i < 1000)
 	{
 		if (i == 0)
-			printf("%lu", j);
+			printf("%.0Lf", j);
 		else if (i == 1)
-			printf(", %lu", k);
+			printf(", %.0Lf", k);
 		else
 		{
 			k += j;
 			j = k - j;
 			if (m < 99)
 			{
-				printf(", %lu", k);
+				printf(", %.0Lf", k);
 			}
 			m++;
 		}
